Last argument to sys_psynch_cvwait() is only 32 bits wide

+++ src/kernel/emulation/linux/psynch/psynch_cvwait.c
@@ -7,7 +7,7 @@
 #include "../simple.h"
 
 long sys_psynch_cvwait(void* cv, uint32_t cvgen, uint32_t cvugen, void* mutex, uint32_t mgen,
-		uint32_t ugen, uint64_t sec, uint64_t usec)
+		uint32_t ugen, uint64_t sec, uint32_t usec)
 {
 	struct psynch_cvwait_args args = {
 		.cv = cv,
